Cleanup credential inheritance tree
