+++ packages-exp/auth-exp/src/core/credentials/index.d.ts
@@ -16,15 +16,13 @@
  */
 
 import * as externs from '@firebase/auth-types-exp';
+import { PhoneOrOauthTokenResponse } from '../../api/authentication/mfa';
+import { IdTokenResponse } from '../../model/id_token';
+import { Auth } from '../../model/auth';
 
-import { Auth } from './auth';
-import { IdTokenResponse } from './id_token';
-import { PhoneOrOauthTokenResponse } from '../api/authentication/mfa';
-
-export interface AuthCredential extends externs.AuthCredential {
-  readonly providerId: externs.ProviderId;
-  readonly signInMethod: externs.SignInMethod;
-  toJSON(): object;
+export abstract class AuthCredential extends externs.AuthCredential {
+  static fromJSON(json: object | string): AuthCredential | null;
+  
   _getIdTokenResponse(auth: Auth): Promise<PhoneOrOauthTokenResponse>;
   _linkToIdToken(auth: Auth, idToken: string): Promise<IdTokenResponse>;
   _matchIdTokenWithUid(auth: Auth, uid: string): Promise<IdTokenResponse>;
