+++ packages-exp/auth-exp/src/core/providers/anonymous.ts
@@ -15,43 +15,8 @@
  * limitations under the License.
  */
 
-import {
-  AuthCredential,
-  ProviderId,
-  SignInMethod,
-  AuthProvider
-} from '@firebase/auth-types-exp';
-import { signUp } from '../../api/authentication/sign_up';
-import { Auth } from '../../model/auth';
-import { IdTokenResponse } from '../../model/id_token';
-import { debugFail } from '../util/assert';
-
-export class AnonymousCredential implements AuthCredential {
-  providerId = ProviderId.ANONYMOUS;
-  signInMethod = SignInMethod.ANONYMOUS;
-
-  toJSON(): never {
-    debugFail('Method not implemented.');
-  }
-
-  static fromJSON(): never {
-    debugFail('Method not implemented');
-  }
-
-  async _getIdTokenResponse(auth: Auth): Promise<IdTokenResponse> {
-    return signUp(auth, {
-      returnSecureToken: true
-    });
-  }
-
-  async _linkToIdToken(_auth: Auth, _idToken: string): Promise<never> {
-    debugFail("Can't link to an anonymous credential");
-  }
-
-  _matchIdTokenWithUid(_auth: Auth, _uid: string): Promise<never> {
-    debugFail('Method not implemented.');
-  }
-}
+import { AuthProvider, ProviderId } from '@firebase/auth-types-exp';
+import { AnonymousCredential } from '../credentials/anonymous';
 
 export class AnonymousProvider implements AuthProvider {
   providerId = ProviderId.ANONYMOUS;
